Move resolve_neurite_types from weights.py to positions.py
The function is only called in positions.py, so colocate it with its
sole caller and remove the cross-module import.
